+++ HISTORY.md
@@ -1,6 +1,7 @@
 unreleased
 ==========
 
+  * Add fast-path for empty entity
   * Shrink size of generated ETags
 
 1.0.1 / 2014-08-24